custom Mask class for faster segmentation results when mapped
