+++ coralnet_toolbox/AutoDistill/QtDeployModel.py
@@ -17,7 +17,8 @@ from PyQt5.QtWidgets import (QApplication, QComboBox, QDialog,
 
 from coralnet_toolbox.QtProgressBar import ProgressBar
 
-from coralnet_toolbox.ResultsProcessor import ResultsProcessor
+from coralnet_toolbox.Results import ResultsProcessor
+from coralnet_toolbox.Results import MapResults
 
 from coralnet_toolbox.utilities import rasterio_open
 from coralnet_toolbox.utilities import rasterio_to_numpy
@@ -452,7 +453,7 @@ class DeployModelDialog(QDialog):
 
         # Make cursor busy
         QApplication.setOverrideCursor(Qt.WaitCursor)
-        
+
         # Start the progress bar
         progress_bar = ProgressBar(self.annotation_window, title="Prediction Workflow")
         progress_bar.show()
@@ -468,10 +469,10 @@ class DeployModelDialog(QDialog):
                 results = self._update_results(results_processor, results, inputs, image_path)
                 results = self._apply_sam(results, image_path)
                 self._process_results(results_processor, results, image_path)
-                
+
                 # Update the progress bar
                 progress_bar.update_progress()
-                
+
         except Exception as e:
             print("An error occurred during prediction:", e)
         finally:
@@ -492,16 +493,16 @@ class DeployModelDialog(QDialog):
         else:
             # Get the work areas
             work_areas_data = raster.get_work_areas_data()
-            
+
         return work_areas_data
-    
+
     def _apply_model(self, inputs):
         """Apply the model to the inputs."""
         # Start the progress bar
         progress_bar = ProgressBar(self.annotation_window, title="Making Predictions")
         progress_bar.show()
         progress_bar.start_progress(len(inputs))
-        
+
         results_list = []
 
         # Process each input separately
@@ -515,16 +516,16 @@ class DeployModelDialog(QDialog):
             # Clean up GPU memory after each prediction
             gc.collect()
             empty_cache()
-                
+
         # Close the progress bar
         progress_bar.finish_progress()
         progress_bar.stop_progress()
         progress_bar.close()
-                
+
         return results_list
 
     def _update_results(self, results_processor, results, inputs, image_path):
-        """Update the results to match Ultralytics format."""           
+        """Update the results to match Ultralytics format."""
         return [results_processor.from_supervision(results,
                                                    inputs,
                                                    image_path,
@@ -535,25 +536,25 @@ class DeployModelDialog(QDialog):
         # Check if SAM model is deployed
         if self.use_sam_dropdown.currentText() != "True":
             return results_list
-        
+
         # Make cursor busy
         QApplication.setOverrideCursor(Qt.WaitCursor)
         progress_bar = ProgressBar(self.annotation_window, title="Predicting with SAM")
         progress_bar.show()
         progress_bar.start_progress(len(results_list))
-        
+
         updated_results = []
-        
+
         for idx, results in enumerate(results_list):
             # Each Results is a list (within the results_list, [[], ]
             if results:
                 # Run it rough the SAM model
                 results = self.sam_dialog.predict_from_results(results, image_path)
                 updated_results.append(results)
-            
+
             # Update the progress bar
             progress_bar.update_progress()
-            
+
         # Make cursor normal
         QApplication.restoreOverrideCursor()
         progress_bar.finish_progress()
@@ -567,15 +568,15 @@ class DeployModelDialog(QDialog):
         # Get the raster object and number of work items
         raster = self.image_window.raster_manager.get_raster(image_path)
         total = raster.count_work_items()
-        
+
         # Get the work areas (if any)
         work_areas = raster.get_work_areas()
-        
+
         # Start the progress bar
         progress_bar = ProgressBar(self.annotation_window, title="Processing Results")
         progress_bar.show()
         progress_bar.start_progress(total)
-        
+
         updated_results = []
 
         for idx, results in enumerate(results_list):
@@ -583,27 +584,30 @@ class DeployModelDialog(QDialog):
             if results:
                 # Update path and names
                 results[0].path = image_path
-                
+
                 # Check if the work area is valid, or the image path is being used
                 if work_areas and self.annotation_window.get_selected_tool() == "work_area":
                     # Map results from work area to the full image
-                    results = results_processor.map_results_from_work_area(results[0], raster, work_areas[idx])
+                    results = MapResults().map_results_from_work_area(results[0], 
+                                                                      raster, 
+                                                                      work_areas[idx],
+                                                                      self.use_sam_dropdown.currentText() == "True")
                 else:
                     results = results[0]
-                    
+
                 # Append the result object (not a list) to the updated results list
                 updated_results.append(results)
-                    
+
                 # Update the index for the next work area
                 idx += 1
                 progress_bar.update_progress()
-        
+
         # Process the Results
         if self.use_sam_dropdown.currentText() == "True":
             results_processor.process_segmentation_results(updated_results)
         else:
             results_processor.process_detection_results(updated_results)
-            
+
         # Close the progress bar
         progress_bar.finish_progress()
         progress_bar.stop_progress()
