+++ coralnet_toolbox/SAM/QtDeployGenerator.py
@@ -15,8 +15,8 @@ from PyQt5.QtWidgets import (QApplication, QComboBox, QDialog, QFormLayout, QHBo
                              QLabel, QMessageBox, QPushButton, QSlider, QSpinBox,
                              QVBoxLayout, QGroupBox)
 
-from coralnet_toolbox.QtWorkArea import WorkArea
-from coralnet_toolbox.ResultsProcessor import ResultsProcessor
+from coralnet_toolbox.Results import ResultsProcessor
+from coralnet_toolbox.Results import MapResults
 
 from coralnet_toolbox.QtProgressBar import ProgressBar
 
@@ -308,9 +308,9 @@ class DeployGeneratorDialog(QDialog):
     def update_task(self):
         """Update the task based on the dropdown selection and handle UI/model effects."""
         self.task = self.use_task_dropdown.currentText()
-        
+
         # Update UI elements based on task
-        if self.task == "segment":            
+        if self.task == "segment":
             # Deactivate model if one is loaded and we're switching to segment task
             if self.loaded_model:
                 self.deactivate_model()
@@ -466,7 +466,7 @@ class DeployGeneratorDialog(QDialog):
 
         # Make cursor busy
         QApplication.setOverrideCursor(Qt.WaitCursor)
-        
+
         # Start the progress bar
         progress_bar = ProgressBar(self.annotation_window, title="Prediction Workflow")
         progress_bar.show()
@@ -481,10 +481,10 @@ class DeployGeneratorDialog(QDialog):
                 results = self._apply_model(inputs)
                 results = self._apply_sam(results, image_path)
                 self._process_results(results_processor, results, image_path)
-                
+
                 # Update the progress bar
                 progress_bar.update_progress()
-                
+
         except Exception as e:
             print("An error occurred during prediction:", e)
         finally:
@@ -505,7 +505,7 @@ class DeployGeneratorDialog(QDialog):
         else:
             # Get the work areas
             work_areas_data = raster.get_work_areas_data()
-            
+
         return work_areas_data
 
     def _apply_model(self, inputs):
@@ -514,12 +514,12 @@ class DeployGeneratorDialog(QDialog):
         self.loaded_model.conf = self.main_window.get_uncertainty_thresh()
         self.loaded_model.iou = self.main_window.get_iou_thresh()
         self.loaded_model.max_det = self.get_max_detections()
-        
+
         # Start the progress bar
         progress_bar = ProgressBar(self.annotation_window, title="Making Predictions")
         progress_bar.show()
         progress_bar.start_progress(len(inputs))
-        
+
         results_list = []
 
         # Process each input separately
@@ -533,12 +533,12 @@ class DeployGeneratorDialog(QDialog):
                 # Clean up GPU memory after each prediction
                 gc.collect()
                 empty_cache()
-                
+
         # Close the progress bar
         progress_bar.finish_progress()
         progress_bar.stop_progress()
         progress_bar.close()
-                
+
         return results_list
 
     def _apply_sam(self, results_list, image_path):
@@ -546,28 +546,28 @@ class DeployGeneratorDialog(QDialog):
         # Check if SAM model is deployed
         if self.use_sam_dropdown.currentText() != "True":
             return results_list
-        
+
         # Update the task to segment
         self.task = 'segment'
-        
+
         # Make cursor busy
         QApplication.setOverrideCursor(Qt.WaitCursor)
         progress_bar = ProgressBar(self.annotation_window, title="Predicting with SAM")
         progress_bar.show()
         progress_bar.start_progress(len(results_list))
-        
+
         updated_results = []
-        
+
         for idx, results in enumerate(results_list):
             # Each Results is a list (within the results_list, [[], ]
             if results:
                 # Run it rough the SAM model
                 results = self.sam_dialog.predict_from_results(results, image_path)
                 updated_results.append(results)
-            
+
             # Update the progress bar
             progress_bar.update_progress()
-            
+
         # Make cursor normal
         QApplication.restoreOverrideCursor()
         progress_bar.finish_progress()
@@ -575,21 +575,21 @@ class DeployGeneratorDialog(QDialog):
         progress_bar.close()
 
         return updated_results
-        
+
     def _process_results(self, results_processor, results_list, image_path):
         """Process the results using the result processor."""
         # Get the raster object and number of work items
         raster = self.image_window.raster_manager.get_raster(image_path)
         total = raster.count_work_items()
-        
+
         # Get the work areas (if any)
         work_areas = raster.get_work_areas()
-        
+
         # Start the progress bar
         progress_bar = ProgressBar(self.annotation_window, title="Processing Results")
         progress_bar.show()
         progress_bar.start_progress(total)
-        
+
         updated_results = []
 
         for idx, results in enumerate(results_list):
@@ -598,27 +598,30 @@ class DeployGeneratorDialog(QDialog):
                 # Update path and names
                 results[0].path = image_path
                 results[0].names = {0: self.class_mapping[0].short_label_code}
-                
+
                 # Check if the work area is valid, or the image path is being used
                 if work_areas and self.annotation_window.get_selected_tool() == "work_area":
                     # Map results from work area to the full image
-                    results = results_processor.map_results_from_work_area(results[0], raster, work_areas[idx])
+                    results = MapResults().map_results_from_work_area(results[0], 
+                                                                      raster, 
+                                                                      work_areas[idx],
+                                                                      self.task == "segment")
                 else:
                     results = results[0]
-                    
+
                 # Append the result object (not a list) to the updated results list
                 updated_results.append(results)
-                    
+
                 # Update the index for the next work area
                 idx += 1
                 progress_bar.update_progress()
-        
+
         # Process the Results
         if self.task == 'segment' or self.use_sam_dropdown.currentText() == "True":
             results_processor.process_segmentation_results(updated_results)
         else:
             results_processor.process_detection_results(updated_results)
-            
+
         # Close the progress bar
         progress_bar.finish_progress()
         progress_bar.stop_progress()
